feat: Create modular Button Component
Refactor the button component to be more modular so that
it can be reused more often. The button now takes in the
following optional props:

* size: PropTypes.oneOf(['inline', 's', 'm', 'l', 'block'])
* color: PropTypes.oneOf(['success', 'warning', 'danger']),
* customStyles: PropTypes.object,

The customStyles optional props is used to pass in a
CSS-in-JS style object to customize the button beyond the
limitations of the size and color props.