refactor: Filter posts in selector
Create a selector in the postsReducer file that takes in
the slice of state (posts object) and a filter string and
subsequently returns an array of post objects filtered by
the category property.

This selector can be used in the mapStateToProps function
for passing a postList to a component in charge of rendering
it.