refactor: Normalize state shape of posts
According to the Redux docs, the state should be treated as
a database and the state should be normalized.

In this app, posts are divided into two entities:

* A key-value (object) collection of post objects called
  byId
* An array of ids that each act as a key for byId called
  allIds

The postsReducer is broken down into two reducers and each
of them are moved into separate files for improved code
re-use and readability.

URL: (https://redux.js.org/recipes/structuring-reducers/normalizing-state-shape)