refactor: Sort and Selector components
Selector component can display an icon alongside the text
for each item in the list. Therefore, selector receive an
options prop that is an array of objects rather than an
array of strings.

The Sort and selector components css files have been
modified to match the look and feel of the rest of the app.