feat: Setup store for handling loaders
New action types and creators have been created to set and
unset the UI states, postsLoading and commentsLoading.

When a server request starts, the set action is dispatched
and once the server returns a response the unset action is
dispatched. This has been set up for both posts and comments.